+++ kadai1/kotaaaa/README.md
@@ -17,8 +17,8 @@
 ```
 $ pwd
 (YOUR_PATH)/gopherdojo-studyroom/kadai1/kotaaaa
-$ go build main.go 
-$ ./main -path="./pic/" -srcExt=".png" -dstExt=".jpg" 
+$ go build -o converter
+$ ./converter -path="./samples/" -srcExt=".jpg" -dstExt=".png"
 ```
 
 ### Help
@@ -33,10 +33,4 @@ Usage of /var/folders/nx/xqljz2y954qbyppfwn4w0tcr0000gn/T/go-build027276676/b001
         変換前の拡張子 (default ".jpg")
 exit status 2
 ```
-
-### TODO 
--  再帰的にファイル取得
-- ユーザ定義型を作ってみる
-- GoDocを生成してみる
-- Go Modulesを使ってみる
