+++ kadai1/kotaaaa/go.mod
@@ -1,0 +1,5 @@
+module github.com/kotaaaa/gopherdojo-studyroom/kadai1/kotaaaa
+
+go 1.14
+
+require golang.org/x/tools v0.1.7 // indirect
